pr requests plus decoding tests
