+++ src/Network/Ethereum/Web3/Encoding/Event.hs
@@ -1,15 +1,27 @@
 {-# LANGUAGE DataKinds              #-}
-{-# LANGUAGE FlexibleContexts       #-}
 {-# LANGUAGE FlexibleInstances      #-}
 {-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE GADTs                  #-}
 {-# LANGUAGE MultiParamTypeClasses  #-}
 {-# LANGUAGE PolyKinds              #-}
 {-# LANGUAGE ScopedTypeVariables    #-}
 {-# LANGUAGE TypeFamilies           #-}
-{-# LANGUAGE TypeInType             #-}
 {-# LANGUAGE TypeOperators          #-}
 {-# LANGUAGE UndecidableInstances   #-}
+
+
+-- |
+-- Module      :  Network.Ethereum.Web3.Encoding.Event
+-- Copyright   :  Alexander Krupenkin 2016
+-- License     :  BSD3
+--
+-- Maintainer  :  mail@akru.me
+-- Stability   :  experimental
+-- Portability :  portable
+--
+-- This module is internal, the purpose is to define helper classes and functions
+-- to assist in event decoding. The user of this library should have no need to use
+-- this directly in application code.
+--
 
 module Network.Ethereum.Web3.Encoding.Event(
     DecodeEvent(..)
@@ -18,22 +30,21 @@ module Network.Ethereum.Web3.Encoding.Event(
   , genericArrayParser
   ) where
 
-import           Data.Kind
-import           Data.Proxy                              (Proxy (..))
-import           Data.Tagged                             (Tagged (..))
 import qualified Data.Text                               as T
 import           Generics.SOP
 import qualified GHC.Generics                            as GHC (Generic)
-import           GHC.TypeLits                            (CmpNat, Nat)
 
 import           Network.Ethereum.Web3.Address           (Address)
 import           Network.Ethereum.Web3.Encoding          (ABIDecode, fromData)
 import           Network.Ethereum.Web3.Encoding.Generic  (GenericABIDecode,
                                                           genericFromData)
+import           Network.Ethereum.Web3.Encoding.Event.Internal
 import           Network.Ethereum.Web3.Encoding.Internal
 import           Network.Ethereum.Web3.Types             (Change (..))
 
-
+-- | Indexed event args come back in as a list of encoded values. 'ArrayParser'
+-- | is used to decode these values so that they can be used to reconstruct the
+-- | entire decoded event.
 class ArrayParser a where
   arrayParser :: [T.Text] -> Maybe a
 
@@ -65,6 +76,7 @@ genericArrayParser = fmap to . arrayParser
 
 data Event i ni = Event i ni
 
+-- | 'parseChange' decodes both the indexed and non-indexed event components.
 parseChange :: ( Generic i
                , Show i
                , Rep i ~ irep
@@ -92,6 +104,9 @@ parseChange change isAnonymous = do
 class IndexedEvent i ni e | e -> i ni where
   isAnonymous :: Proxy e -> Bool
 
+-- | 'CombineChange' is a class which indicates that given event components of types 'i'
+-- | and 'ni', we can construct an event of type 'e'. The functional dependency is valid
+-- | becasue of how the template haskell generates the event types.
 class CombineChange i ni e | e -> i ni where
   combineChange :: i -> ni -> e
 
@@ -117,7 +132,6 @@ instance ( Generic i
 class DecodeEvent i ni e | e -> i ni where
   decodeEvent :: Change -> Maybe e
 
-
 instance ( IndexedEvent i ni e
          , Generic i
          , Rep i ~ SOP I '[hli]
@@ -134,109 +148,4 @@ instance ( IndexedEvent i ni e
       let anonymous = isAnonymous (Proxy :: Proxy e)
       (Event i ni :: Event i ni) <- parseChange change anonymous
       return $ combineChange i ni
-
---------------------------------------------------------------------------------
--- Event Parsing Internals
---------------------------------------------------------------------------------
-
-data HList :: [*] -> * where
-  HNil :: HList '[]
-  (:<) :: a -> HList as -> HList (a : as)
-
-infixr 0 :<
-
--- | Generic representation to HList representation
-class HListRep a xs | a -> xs, a -> xs where
-  toHList :: a -> HList xs
-  fromHList :: HList xs -> a
-
-instance HListRep (NP I '[]) '[] where
-  toHList _ = HNil
-  fromHList _ = Nil
-
-instance HListRep (NP I as) as => HListRep (NP I (a:as)) (a:as) where
-  toHList (I a :* rest) = a :< toHList rest
-  fromHList (a :< rest) = I a :* fromHList rest
-
-instance HListRep (NP f as') as => HListRep (SOP f '[as']) as where
-  toHList (SOP (Z rep)) = toHList rep
-  fromHList = SOP . Z . fromHList
-
--- | Sort a Tagged HList
-class Sort (xs :: [*]) where
-  type Sort' xs :: [*]
-  sort :: HList xs -> HList (Sort' xs)
-
-instance Sort '[] where
-  type Sort' '[] = '[]
-  sort HNil = HNil
-
-instance (Sort xs, Insert x (Sort' xs)) => Sort (x : xs) where
-  type Sort' (x : xs) = Insert' x (Sort' xs)
-  sort (x :< xs) = insert x (sort xs)
-
-class Insert (x :: *) (xs :: [*]) where
-  type Insert' x xs :: [*]
-  insert :: x -> HList xs -> HList (Insert' x xs)
-
-instance Insert x '[] where
-  type Insert' x '[] = '[x]
-  insert x HNil = x :< HNil
-
-instance InsertCmp (CmpNat n m) (Tagged n x) (Tagged m y) ys => Insert (Tagged n x) (Tagged m y : ys) where
-  type Insert' (Tagged n x) (Tagged m y : ys) = InsertCmp' (CmpNat n m) (Tagged n x) (Tagged m y) ys
-  insert (x :: Tagged n x) ((y :: Tagged m y) :< ys) = insertCmp (Proxy :: Proxy (CmpNat n m)) x y ys
-
-class InsertCmp (b :: Ordering) (x :: *) (y :: *) (ys :: [*]) where
-  type InsertCmp' b x y ys :: [*]
-  insertCmp :: Proxy (b :: Ordering) -> x -> y -> HList ys -> HList (InsertCmp' b x y ys)
-
-instance InsertCmp 'LT x y ys where
-  type InsertCmp' 'LT x y ys = x : (y : ys)
-  insertCmp _ x y ys = x :< y :< ys
-
-instance Insert x ys => InsertCmp 'GT x y ys where
-  type InsertCmp' 'GT x y ys = y : Insert' x ys
-  insertCmp _ x y ys = y :< insert x ys
-
--- | Unwrap all the Tagged items in an HList
-class UnTag t where
-  type UnTag' t :: [*]
-  unTag :: HList t -> HList (UnTag' t)
-
-instance UnTag '[] where
-  type UnTag' '[] = '[]
-  unTag a = a
-
-instance UnTag ts => UnTag (Tagged n a : ts) where
-  type UnTag' (Tagged n a : ts) = a : UnTag' ts
-  unTag (Tagged a :< ts) = a :< unTag ts
-
-class HListMerge (as :: [*]) (bs :: [*]) where
-  type Concat as bs :: [*]
-  mergeHList :: HList as -> HList bs -> HList (Concat as bs)
-
-instance HListMerge '[] bs where
-  type Concat '[] bs = bs
-  mergeHList _ bs = bs
-
-instance HListMerge as bs => HListMerge (a : as) bs where
-  type Concat (a : as) bs = a : Concat as bs
-  mergeHList (a :< as) bs = a :< mergeHList as bs
-
-class HListMergeSort as bs where
-  type MergeSort' as bs :: [*]
-  mergeSortHList :: HList as -> HList bs -> HList (MergeSort' as bs)
-
-instance (HListMerge as bs, Concat as bs ~ cs, Sort cs, Sort' cs ~ cs') => HListMergeSort as bs where
-  type MergeSort' as bs = Sort' (Concat as bs)
-  mergeSortHList as bs = sort $ (mergeHList as bs :: HList cs)
-
-class MergeIndexedArguments as bs where
-  type MergeIndexedArguments' as bs :: [*]
-  mergeIndexedArguments :: HList as -> HList bs -> HList (MergeIndexedArguments' as bs)
-
-instance (HListMergeSort as bs, MergeSort' as bs ~ cs, UnTag cs, UnTag cs' ~ ds) => MergeIndexedArguments as bs where
-  type MergeIndexedArguments' as bs = (UnTag' (MergeSort' as bs))
-  mergeIndexedArguments as bs = unTag $ mergeSortHList as bs
