+++ src/Network/Ethereum/Web3/JsonAbi.hs
@@ -1,5 +1,5 @@
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell   #-}
+
 -- |
 -- Module      :  Network.Ethereum.Web3.JsonAbi
 -- Copyright   :  Alexander Krupenkin 2016
